feat: 통계 기능 마무리
